+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 - CookieUtil::parseDate to create a date from cookie date string
 
+### Fixed
+
+- Fix curl command of CurlFormatter when there is an user-agent header   
+
 ## 1.5.0 - 2017-02-14
 
 ### Added